invert `BeaconChain` dependency
